# Conflicts:
#	application/database/changelog/versions/1.0.0/main.xml

+++ application/schr-scheduler/patient-consents-topic-import-service/patient-consents-topic-domain/src/main/java/model/InterventionDetails.java
@@ -3,20 +3,20 @@ package model;
 import java.util.List;
 
 public class InterventionDetails {
-    private List<Integer> medInterventionId;
+    private List<Long> medInterventionId;
 
     public InterventionDetails() {
     }
 
-    public InterventionDetails(List<Integer> interventionDetails) {
+    public InterventionDetails(List<Long> interventionDetails) {
         this.medInterventionId = interventionDetails;
     }
 
-    public List<Integer> getMedInterventionId() {
+    public List<Long> getMedInterventionId() {
         return medInterventionId;
     }
 
-    public void setMedInterventionId(List<Integer> medInterventionId) {
+    public void setMedInterventionId(List<Long> medInterventionId) {
         this.medInterventionId = medInterventionId;
     }
 }
